requirements.txt
Django~=4.0
Pillow==8.4.0
djangorestframework==3.13.0